v1/requirements.txt
# Enhanced QnA Agent System - MVP Requirements

# Core dependencies
streamlit>=1.28.0
flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
python-engineio>=4.7.0

# Database
psycopg2-binary>=2.9.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# AI/ML
langchain>=0.0.300
langchain-openai>=0.0.2
openai>=1.0.0

# HTTP requests
requests>=2.31.0

# Environment and configuration
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0 